Moved is_scalar
