+++ include/boost/simd/meta/is_scalar.hpp
